melhorando o codigo e fazendo alteracoes

+++ TheatricalPlayersRefactoringKata/Calculos_Valores_Base_Peça/ICalculoValoresBasePeça.cs
@@ -2,5 +2,5 @@
 
 public interface ICalculoValoresBasePeça
 {
-    int CalculaValoresBase(Performance perf, Play play);
+    decimal CalculaValoresBase(Performance perf, Play play);
 }
